+++ src/actions/types.ts
@@ -1,12 +1,5 @@
 export enum dataActionTypes {
     ADD_TODO = 'ADD_TODO',
     DEL_TODO = 'DEL_TODO',
-    UPDATE_TODO = 'UPDATE_TODO',
-    SET_VIEW = 'SET_VIEW'
-}
-
-export enum viewActionTypes {
-    INVERSE_VIEW = 'INVERSE_VIEW',
-    SET_CLOSED = 'SET_CLOSED',
-    SET_OPENED = 'SET_OPENED'
+    UPDATE_TODO = 'UPDATE_TODO'
 }