+++ src/reducers/NoteReducer.ts
@@ -1,62 +1,68 @@
-import { dataActionTypes, viewActionTypes } from '../actions/types';
+import { dataActionTypes} from '../actions/types';
 import NoteColors from '../constans/NoteColors';
-import ViewReducer, { IViewsState } from './NoteViewReducer';
 import DataReducer, {IDataState} from './NoteDataReducer';
 import { combineReducers } from 'redux';
 
-export interface INoteState {
-    id: number;
-    dataState: IDataState;
-    viewState: IViewsState;
+export interface InotesByID {
+    [index: number]: IDataState;
 }
 
-export default (state: INoteState[] = [], action: any): INoteState[] => {
+export interface INoteAppState {
+    notesID: number[]
+    notesByID: InotesByID
+}
+
+const notesIDReducer = (state: number[] = [], action: any): number[] => {
     switch (action.type) {
         case dataActionTypes.ADD_TODO: {
             return [
                 ...state,
-                {
-                    id: action.id,
-                    dataState: DataReducer(undefined, action),
-                    viewState: ViewReducer(undefined, action)
-                }
-
+                action.id
             ];
         }
         case dataActionTypes.UPDATE_TODO: {
-            return state.map( note => {
-                if (note.id === action.id) {
-                    return {
-                        ...note,
-                        dataState: DataReducer(note.dataState, action)
-                    };
-                }
-                return note;
-            });
+            return [...state];
         }
         case dataActionTypes.DEL_TODO: {
             let stateAfter = [];
-            state.forEach(note => {
-                if (note.id !== action.id) {
-                    stateAfter = [ ...stateAfter, note ];
+            state.forEach(id => {
+                if (id !== action.id) {
+                    stateAfter = [ ...stateAfter, id ];
                 }
             });
-
             return stateAfter;
         }
-        case dataActionTypes.SET_VIEW: {
-            return state.map( note => {
-                if (note.id === action.id) {
-                    return {
-                        ...note,
-                        viewState: ViewReducer(note.viewState, {type: action.viewType})
-                    };
-                }
-                return note;
-            });
+        default: {
+            return state;
+        }
+ }
+};
+
+const notesByIDReducer = (state: InotesByID = [], action: any): InotesByID => {
+    switch (action.type) {
+        case dataActionTypes.ADD_TODO: {
+            let nextState: InotesByID = Object.assign( {}, state );
+            nextState[action.id] = DataReducer( undefined, action );
+            return nextState;
+        }
+        case dataActionTypes.UPDATE_TODO: {
+            let nextState: InotesByID = Object.assign( {}, state);
+            nextState[action.id] = DataReducer( state[action.id], action );
+            return nextState;
+        }
+        case dataActionTypes.DEL_TODO: {
+            let nextState: InotesByID = Object.assign( {}, state);
+            delete nextState[action.id];
+            return nextState;
+
         }
         default: {
             return state;
         }
     }
 };
+
+export default combineReducers({
+    notesID: notesIDReducer,
+    notesByID: notesByIDReducer
+});