+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.0.6] - 2020-03-14
+
+### Changes
+
+    - NoteView reducer deleted.
+    - noteView actions deleted.
+    - note view state types deleted.
+    - Reducers tree refactored ( 
+        - top level state consist with notesID array, and notesById.
+     )
+    - components refactored ( note, noteList, addButton )
+
+### Added
+
+    - notesId state added.
+    - notesByID state added.
+
 ## [0.0.5] - 2020-03-13
 
 ### Changes